+++ source/extensions/isaacsim.core.nodes/python/tests/test_core_nodes.py
@@ -61,12 +61,6 @@ class TestCoreNodes(omni.kit.test.AsyncTestCase):
 
     # ----------------------------------------------------------------------
     async def test_physics_num_steps(self):
-        carb.settings.get_settings().set_bool("/app/runLoops/main/rateLimitEnabled", True)
-        carb.settings.get_settings().set_int("/app/runLoops/main/rateLimitFrequency", 60)
-        carb.settings.get_settings().set_int("/persistent/simulation/minFrameRate", 60)
-        omni.timeline.get_timeline_interface().set_target_framerate(60)
-        omni.timeline.get_timeline_interface().set_time_codes_per_second(60)
-
         steps = self._core_nodes.get_physics_num_steps()
         self.assertEqual(steps, 0)
         self._timeline.play()