+++ source/extensions/isaacsim.robot_setup.assembler/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [3.1.1] - 2025-10-18
+### Changed
+- Remove extra carb settings from tests
 
 ## [3.1.0] - 2025-10-17
 ### Changed